+++ erogenousbeef/bigreactors/client/gui/GuiReactorControlRod.java
@@ -1,7 +1,12 @@
 package erogenousbeef.bigreactors.client.gui;
 
+import java.awt.event.KeyEvent;
+
+import org.lwjgl.input.Keyboard;
+
 import cpw.mods.fml.common.network.PacketDispatcher;
 import net.minecraft.client.gui.GuiButton;
+import net.minecraft.client.gui.GuiTextField;
 import net.minecraft.inventory.Container;
 import net.minecraft.util.ResourceLocation;
 import erogenousbeef.bigreactors.common.BigReactors;
@@ -19,9 +24,14 @@ public class GuiReactorControlRod extends BeefGuiBase {
 	BeefGuiLabel fuelString;
 	BeefGuiLabel wasteString;
 	BeefGuiLabel rodStatus;
+	BeefGuiLabel rodNameLabel;
 
 	GuiButton rodInsertBtn;
 	GuiButton rodRetractBtn;
+	GuiButton setNameBtn;
+
+    private GuiTextField rodName;
+
 	
 	public GuiReactorControlRod(Container c, TileEntityReactorControlRod controlRod) {
 		super(c);
@@ -42,8 +52,22 @@ public class GuiReactorControlRod extends BeefGuiBase {
 		int leftX = 4;
 		int topY = 4;
 		
+		Keyboard.enableRepeatEvents(true);
+		
 		titleString = new BeefGuiLabel(this, "Reactor Control Rod", leftX, topY);
 		topY += titleString.getHeight() + 8;
+		
+		rodNameLabel = new BeefGuiLabel(this, "Name:", leftX, topY + 6);
+		
+		rodName = new GuiTextField(fontRenderer, guiLeft + leftX*2 + rodNameLabel.getWidth(), guiTop + topY, 100, 20);
+		rodName.setCanLoseFocus(true);
+		rodName.setMaxStringLength(32);
+		rodName.setText(entity.getName());
+		rodName.setEnabled(true);
+		
+		setNameBtn = new GuiButton(2, guiLeft + 140, guiTop + topY, 30, 20, "Set");
+		setNameBtn.enabled = false;
+		topY += 28;
 		
 		heatString = new BeefGuiLabel(this, "Heat: ??? C", leftX, topY);
 		topY += heatString.getHeight() + 8;
@@ -68,9 +92,19 @@ public class GuiReactorControlRod extends BeefGuiBase {
 		registerControl(fuelString);
 		registerControl(wasteString);
 		registerControl(rodStatus);
+		registerControl(rodNameLabel);
+		
+		registerControl(rodName);
 		
 		buttonList.add(rodRetractBtn);
 		buttonList.add(rodInsertBtn);
+		buttonList.add(setNameBtn);
+	}
+	
+	@Override
+	public void onGuiClosed() {
+		super.onGuiClosed();
+		Keyboard.enableRepeatEvents(false);
 	}
 	
 	@Override
@@ -89,6 +123,8 @@ public class GuiReactorControlRod extends BeefGuiBase {
 			rodInsertBtn.enabled = false;
 			rodRetractBtn.enabled = false;
 		}
+		
+		this.setNameBtn.enabled = !entity.getName().equals(this.rodName.getText());
 	}
 	
 	@Override
@@ -99,6 +135,11 @@ public class GuiReactorControlRod extends BeefGuiBase {
 		case 0:
 			btnCmd = "rodRetract";
 			break;
+		case 2:
+			PacketDispatcher.sendPacketToServer(PacketWrapper.createPacket(BigReactors.CHANNEL, Packets.ControlRodSetName,
+					new Object[] { entity.xCoord, entity.yCoord, entity.zCoord, this.rodName.getText() }));
+			this.rodName.setFocused(false);
+			return;
 		case 1:
 		default:
 			btnCmd = "rodInsert";
@@ -108,4 +149,30 @@ public class GuiReactorControlRod extends BeefGuiBase {
 		PacketDispatcher.sendPacketToServer(PacketWrapper.createPacket(BigReactors.CHANNEL, Packets.BeefGuiButtonPress,
 				new Object[] { entity.xCoord, entity.yCoord, entity.zCoord, btnCmd }));
 	}
+	
+	@Override
+	protected void keyTyped(char inputChar, int keyCode) {
+        if (keyCode == Keyboard.KEY_ESCAPE ||
+        		(!this.rodName.isFocused() && keyCode == this.mc.gameSettings.keyBindInventory.keyCode)) {
+            this.mc.thePlayer.closeScreen();
+        }
+
+		this.rodName.textboxKeyTyped(inputChar, keyCode);
+		
+		if(keyCode == Keyboard.KEY_TAB) {
+			// Tab
+			if(this.rodName.isFocused()) {
+				this.rodName.setFocused(false);
+			}
+			else {
+				this.rodName.setFocused(true);
+			}
+		}
+		
+		if(keyCode == Keyboard.KEY_RETURN) {
+			// Return/enter
+			this.actionPerformed((GuiButton)this.buttonList.get(2));
+		}
+	}
+	
 }