+++ erogenousbeef/bigreactors/gui/controls/grab/IBeefGuiGrabbable.java
@@ -1,0 +1,9 @@
+package erogenousbeef.bigreactors.gui.controls.grab;
+
+import net.minecraft.util.Icon;
+
+public interface IBeefGuiGrabbable {
+
+	public Icon getIcon();
+	public String getName();
+}